refactor: rename packages from 'cep' to 'zipcode'
