+++ internal/dependencies/dependencies.go
@@ -1,9 +1,9 @@
 package dependencies
 
 import (
-	"luizalabs-technical-test/internal/features/cep"
 	"luizalabs-technical-test/internal/features/health"
 	"luizalabs-technical-test/internal/features/swagger"
+	"luizalabs-technical-test/internal/features/zipcode"
 	"luizalabs-technical-test/pkg/http"
 
 	netHttp "net/http"
@@ -15,10 +15,10 @@ import (
 func Load() []func(*gin.RouterGroup) {
 	httpClient := http.NewClient(&netHttp.Client{})
 
-	// cep feature
-	cepRep := cep.NewRepository(httpClient)
-	cepSrv := cep.NewService(cepRep)
-	cepHandler := cep.NewHandler(cepSrv)
+	// zipcode feature
+	zipCodeRep := zipcode.NewRepository(httpClient)
+	zipCodeSrv := zipcode.NewService(zipCodeRep)
+	zipCodeHandler := zipcode.NewHandler(zipCodeSrv)
 
 	// health feature
 	healthHandler := health.NewHandler()
@@ -29,6 +29,6 @@ func Load() []func(*gin.RouterGroup) {
 	return []func(*gin.RouterGroup){
 		swaggerHandler.Register,
 		healthHandler.Register,
-		cepHandler.Register,
+		zipCodeHandler.Register,
 	}
 }
